initial commit: Create website structure with django-admin
The command 'django-admin startproject <name>' creates all the files in this
commit. The name of the root directory was appended a '_project' in order to
be differentiated from the child directory where the website code lives.

The server can be run at this point by running the command
'python3 manage.py runserver' and the initial page can be accessed at localhost:8000

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>