lecture_26: Create templates directory and custom home page
The html pages in Django live in a directory called 'templates/<app_name>'
directory. They are referenced in 'views.py' file, being called by a function
that, as a good practice, has the same name as the html file.

This function in views.py is called by the 'urls.py' file, when the path
inside this file routs the request to the function inside 'views.py'.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>