lecture 28: use form data
This commit inserts code to make the password consume the options provided
by the user in the home.html page.

The changes are all in views.py, inside the password function and makes use
of the data that is contained inside the 'request' parameter, coming to
password() from the form in home.html.

For further information on the request object (HttpRequest object), please
refer to https://docs.djangoproject.com/en/3.1/ref/request-response/

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>